requirements.txt
Django==3.2.9
django_filter==21.1
djangorestframework==3.12.4
